Create README.md
Based on what I have learned, the modelling of a physical phenomenon requires knowledge of mathematics, physics and programming. This is why I applied for a Master's degree in Mathematical Sciences at the African Institute of Mathematical Sciences (AIMS). Before coming at AIMS, I defended a Master's thesis in Atmospheric Sciences on weather forecasting from the Weather Research and forecasting model. I would now like to commit myself to a thesis in physics or in physics-mathematics, in a framework where I can work on my project by collaborating with other scientists, while participating in the activities of the laboratory and the University or for a work in industry which will allow me to acquire the bases necessary to create my own. I strongly believe that working with other people from different backgrounds and cultures will allow me to boost my limits as is the case at AIMS. Although my work is not always of a high standard, I learn from my mistakes and those of others and improve myself every day.